[FIX] Fixing warning in file_mapping.h.
git-svn-id: http://svn.seqan.de/seqan/trunk@13431 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ core/include/seqan/file/file_mapping.h
@@ -358,8 +358,8 @@ template <typename TSpec, typename TPos, typename TSize>
 inline void *
 mapFileSegment(FileMapping<TSpec> &mapping, TPos fileOfs, TSize size, FileMappingMode mode)
 {
-    SEQAN_ASSERT_EQ(OPEN_RDONLY, MAP_RDONLY);
-    SEQAN_ASSERT_EQ(OPEN_WRONLY, MAP_WRONLY);
+    SEQAN_ASSERT_EQ((int)OPEN_RDONLY, (int)MAP_RDONLY);
+    SEQAN_ASSERT_EQ((int)OPEN_WRONLY, (int)MAP_WRONLY);
 
     void *addr;
     if (size == 0)
